requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_tflow_select=2.1.0=gpu
absl-py=0.6.1=py36_0
astor=0.7.1=py36_0
blas=1.0=mkl
bunch=1.0.1=py_1
bzip2=1.0.6=h14c3975_5
c-ares=1.15.0=h7b6447c_1
ca-certificates=2018.10.15=ha4d7672_0
cairo=1.14.12=h8948797_3
cairocffi=0.8.0=py36_1000
certifi=2018.10.15=py36_1000
cffi=1.11.5=py36h5e8e0c9_1
cloudpickle=0.6.1=py36_0
cudatoolkit=9.2=0
cudnn=7.2.1=cuda9.2_0
cupti=9.2.148=0
cycler=0.10.0=py36_0
dask-core=0.20.2=py36_0
dbus=1.13.2=h714fa37_1
decorator=4.3.0=py36_0
expat=2.2.6=he6710b0_0
ffmpeg=4.0=hcdf2ecd_0
fontconfig=2.13.0=h9420a91_0
freeglut=3.0.0=hf484d3e_5
freetype=2.9.1=h8a8886c_1
gast=0.2.0=py36_0
glib=2.56.2=hd408876_0
graphite2=1.3.12=h23475e2_2
grpcio=1.14.1=py36h9ba97e2_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
h5py=2.8.0=py36h989c5e5_3
harfbuzz=1.8.8=hffaf4a1_0
hdf5=1.10.2=hba1933b_1
icu=58.2=h9c2bf20_1
imageio=2.4.1=py36_0
imgaug=0.2.6=0
intel-openmp=2019.1=144
jasper=2.0.14=h07fcdf6_1
jpeg=9b=h024ee3a_2
keras=2.2.4=0
keras-applications=1.0.6=py36_0
keras-base=2.2.4=py36_0
keras-preprocessing=1.0.5=py36_0
kiwisolver=1.0.1=py36hf484d3e_0
libedit=3.1.20170329=h6b74fdf_2
libffi=3.2.1=hd88cf55_4
libgcc-ng=8.2.0=hdf63c60_1
libgfortran-ng=7.3.0=hdf63c60_0
libglu=9.0.0=hf484d3e_1
libopencv=3.4.2=hb342d67_1
libopus=1.3=h7b6447c_0
libpng=1.6.35=hbc83047_0
libprotobuf=3.6.1=hd408876_0
libstdcxx-ng=8.2.0=hdf63c60_1
libtiff=4.0.9=he85c1e1_2
libuuid=1.0.3=h1bed415_2
libvpx=1.7.0=h439df22_0
libxcb=1.13=h1bed415_1
libxml2=2.9.8=h26e45fe_1
markdown=3.0.1=py36_0
matplotlib=3.0.1=py36h5429711_0
mkl=2018.0.3=1
mkl_fft=1.0.6=py36h7dd41cf_0
mkl_random=1.0.1=py36h4414c95_1
ncurses=6.1=hf484d3e_0
networkx=2.2=py36_1
nltk=3.3.0=py36_0
numpy=1.15.4=py36h1d66e8a_0
numpy-base=1.15.4=py36h81de0dd_0
olefile=0.46=py36_0
opencv=3.4.2=py36h6fd60c2_1
openssl=1.0.2p=h470a237_1
pandas=0.23.4=py36h04863e7_0
pcre=8.42=h439df22_0
pillow=5.3.0=py36h34e0f95_0
pip=18.1=py36_0
pixman=0.34.0=h470a237_3
protobuf=3.6.1=py36he6710b0_0
py-opencv=3.4.2=py36hb342d67_1
pycparser=2.19=py_0
pyparsing=2.3.0=py36_0
pyqt=5.9.2=py36h05f1152_2
python=3.6.6=h6e4f718_2
python-dateutil=2.7.5=py36_0
pytz=2018.7=py36_0
pywavelets=1.0.1=py36hdd07704_0
pyyaml=3.13=py36h14c3975_0
qt=5.9.6=h8703b6f_2
readline=7.0=h7b6447c_5
scikit-image=0.14.0=py36hf484d3e_1
scipy=1.1.0=py36hfa4b5c9_1
setuptools=40.6.2=py36_0
sip=4.19.8=py36hf484d3e_0
six=1.11.0=py36_1
sqlite=3.25.3=h7b6447c_0
tensorboard=1.12.0=py36hf484d3e_0
tensorflow=1.12.0=gpu_py36he74679b_0
tensorflow-base=1.12.0=gpu_py36had579c0_0
tensorflow-gpu=1.12.0=h0d30ee6_0
termcolor=1.1.0=py36_1
tk=8.6.8=hbc83047_0
toolz=0.9.0=py36_0
tornado=5.1.1=py36h7b6447c_0
tqdm=4.28.1=py36h28b3542_0
werkzeug=0.14.1=py36_0
wheel=0.32.3=py36_0
xz=5.2.4=h14c3975_4
yaml=0.1.7=had09818_2
zlib=1.2.11=h7b6447c_3
